--- packages/@aws-cdk-testing/framework-integ/test/core/test/integ.enable-additional-metadata.js.snapshot/EnableTelemtryStack.assets.json
@@ -1,19 +1,0 @@
-{
-  "version": "39.0.0",
-  "files": {
-    "b6503dca5951e510ce013d2525009c961587cc7267ceff8ecacb15d7fb7db461": {
-      "source": {
-        "path": "EnableTelemtryStack.template.json",
-        "packaging": "file"
-      },
-      "destinations": {
-        "current_account-current_region": {
-          "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "b6503dca5951e510ce013d2525009c961587cc7267ceff8ecacb15d7fb7db461.json",
-          "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
-        }
-      }
-    }
-  },
-  "dockerImages": {}
-}
