--- packages/@aws-cdk-testing/framework-integ/test/core/test/integ.enable-additional-metadata.js.snapshot/cdk.out
@@ -1,1 +1,0 @@
-{"version":"39.0.0"}
