--- packages/@aws-cdk-testing/framework-integ/test/core/test/integ.enable-additional-metadata.js.snapshot/integ.json
@@ -1,12 +1,0 @@
-{
-  "version": "39.0.0",
-  "testCases": {
-    "Enable Additional Metadata/DefaultTest": {
-      "stacks": [
-        "EnableTelemtryStack"
-      ],
-      "assertionStack": "Enable Additional Metadata/DefaultTest/DeployAssert",
-      "assertionStackName": "EnableAdditionalMetadataDefaultTestDeployAssert3F14AD57"
-    }
-  }
-}
