--- packages/@aws-cdk-testing/framework-integ/test/core/test/integ.enable-additional-metadata.js.snapshot/manifest.json
@@ -1,184 +1,0 @@
-{
-  "version": "39.0.0",
-  "artifacts": {
-    "EnableTelemtryStack.assets": {
-      "type": "cdk:asset-manifest",
-      "properties": {
-        "file": "EnableTelemtryStack.assets.json",
-        "requiresBootstrapStackVersion": 6,
-        "bootstrapStackVersionSsmParameter": "/cdk-bootstrap/hnb659fds/version"
-      }
-    },
-    "EnableTelemtryStack": {
-      "type": "aws:cloudformation:stack",
-      "environment": "aws://unknown-account/unknown-region",
-      "properties": {
-        "templateFile": "EnableTelemtryStack.template.json",
-        "terminationProtection": false,
-        "validateOnSynth": false,
-        "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-deploy-role-${AWS::AccountId}-${AWS::Region}",
-        "cloudFormationExecutionRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-cfn-exec-role-${AWS::AccountId}-${AWS::Region}",
-        "stackTemplateAssetObjectUrl": "s3://cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}/b6503dca5951e510ce013d2525009c961587cc7267ceff8ecacb15d7fb7db461.json",
-        "requiresBootstrapStackVersion": 6,
-        "bootstrapStackVersionSsmParameter": "/cdk-bootstrap/hnb659fds/version",
-        "additionalDependencies": [
-          "EnableTelemtryStack.assets"
-        ],
-        "lookupRole": {
-          "arn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-lookup-role-${AWS::AccountId}-${AWS::Region}",
-          "requiresBootstrapStackVersion": 8,
-          "bootstrapStackVersionSsmParameter": "/cdk-bootstrap/hnb659fds/version"
-        }
-      },
-      "dependencies": [
-        "EnableTelemtryStack.assets"
-      ],
-      "metadata": {
-        "/EnableTelemtryStack/01234test": [
-          {
-            "type": "aws:cdk:analytics:construct",
-            "data": {
-              "visibilityTimeout": {
-                "amount": 300,
-                "unit": {
-                  "label": "seconds",
-                  "isoLabel": "S",
-                  "inMillis": 1000
-                }
-              }
-            }
-          }
-        ],
-        "/EnableTelemtryStack/01234test/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "01234test13C610BE"
-          }
-        ],
-        "/EnableTelemtryStack/MyFunction": [
-          {
-            "type": "aws:cdk:analytics:construct",
-            "data": {
-              "runtime": {
-                "name": "python3.8",
-                "supportsInlineCode": true,
-                "family": 2,
-                "isVariable": false,
-                "bundlingDockerImage": {
-                  "image": "public.ecr.aws/sam/build-python3.8"
-                },
-                "bundlingImage": {
-                  "image": "public.ecr.aws/sam/build-python3.8"
-                },
-                "supportsCodeGuruProfiling": true,
-                "supportsSnapStart": false
-              },
-              "handler": "index.handler",
-              "code": {
-                "code": "def handler(event, context):\n\tprint('The function has been invoked.')",
-                "isInline": true
-              }
-            }
-          }
-        ],
-        "/EnableTelemtryStack/MyFunction/ServiceRole/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "MyFunctionServiceRole3C357FF2"
-          }
-        ],
-        "/EnableTelemtryStack/MyFunction/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "MyFunction3BAA72D1"
-          }
-        ],
-        "/EnableTelemtryStack/BootstrapVersion": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "BootstrapVersion"
-          }
-        ],
-        "/EnableTelemtryStack/CheckBootstrapVersion": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "CheckBootstrapVersion"
-          }
-        ],
-        "MyBucketF68F3FF0": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "MyBucketF68F3FF0",
-            "trace": [
-              "!!DESTRUCTIVE_CHANGES: WILL_DESTROY"
-            ]
-          }
-        ],
-        "CDKMetadata": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "CDKMetadata",
-            "trace": [
-              "!!DESTRUCTIVE_CHANGES: WILL_DESTROY"
-            ]
-          }
-        ]
-      },
-      "displayName": "EnableTelemtryStack"
-    },
-    "EnableAdditionalMetadataDefaultTestDeployAssert3F14AD57.assets": {
-      "type": "cdk:asset-manifest",
-      "properties": {
-        "file": "EnableAdditionalMetadataDefaultTestDeployAssert3F14AD57.assets.json",
-        "requiresBootstrapStackVersion": 6,
-        "bootstrapStackVersionSsmParameter": "/cdk-bootstrap/hnb659fds/version"
-      }
-    },
-    "EnableAdditionalMetadataDefaultTestDeployAssert3F14AD57": {
-      "type": "aws:cloudformation:stack",
-      "environment": "aws://unknown-account/unknown-region",
-      "properties": {
-        "templateFile": "EnableAdditionalMetadataDefaultTestDeployAssert3F14AD57.template.json",
-        "terminationProtection": false,
-        "validateOnSynth": false,
-        "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-deploy-role-${AWS::AccountId}-${AWS::Region}",
-        "cloudFormationExecutionRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-cfn-exec-role-${AWS::AccountId}-${AWS::Region}",
-        "stackTemplateAssetObjectUrl": "s3://cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}/21fbb51d7b23f6a6c262b46a9caee79d744a3ac019fd45422d988b96d44b2a22.json",
-        "requiresBootstrapStackVersion": 6,
-        "bootstrapStackVersionSsmParameter": "/cdk-bootstrap/hnb659fds/version",
-        "additionalDependencies": [
-          "EnableAdditionalMetadataDefaultTestDeployAssert3F14AD57.assets"
-        ],
-        "lookupRole": {
-          "arn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-lookup-role-${AWS::AccountId}-${AWS::Region}",
-          "requiresBootstrapStackVersion": 8,
-          "bootstrapStackVersionSsmParameter": "/cdk-bootstrap/hnb659fds/version"
-        }
-      },
-      "dependencies": [
-        "EnableAdditionalMetadataDefaultTestDeployAssert3F14AD57.assets"
-      ],
-      "metadata": {
-        "/Enable Additional Metadata/DefaultTest/DeployAssert/BootstrapVersion": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "BootstrapVersion"
-          }
-        ],
-        "/Enable Additional Metadata/DefaultTest/DeployAssert/CheckBootstrapVersion": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "CheckBootstrapVersion"
-          }
-        ]
-      },
-      "displayName": "Enable Additional Metadata/DefaultTest/DeployAssert"
-    },
-    "Tree": {
-      "type": "cdk:tree",
-      "properties": {
-        "file": "tree.json"
-      }
-    }
-  }
-}
